Phase 3: the guitar becomes a band member - audio looper
The riskiest subsystem, landed on the Phase 2 capture machinery:

Firmware:
- capture.h AudioRing: always-on rolling input buffer (36s, one bar of
  margin over the largest window) written per-sample in the callback;
  bar ANCHORS {tick, sample_pos} recorded at every bar line make the
  tick<->sample mapping exact by construction; envelope follower feeds
  the guitarist's activity pulse
- audio_track.h GranulePool: tempo lock carves the 42MB pool into
  1-bar granules (fragmentation impossible - any freed bar fits any
  loop); alloc is exact; undo/delete FREE memory. Corrected SPEC
  capacity table (draft double-counted bytes/bar): ~114 bars @60BPM,
  ~229 @120, ~344 @180
- CopyJob: incremental ring->granules copy (8192 samples/main-loop
  pass, ~190ms worst case) with lap-detection abort; granules land in
  LOOP-POSITION order (global bar % bars - the same phase-preserving
  rule as MIDI capture); waveform peaks (24/bar) computed during the
  copy for free
- Tempo lock state machine: locks on first committed audio capture,
  unlocks when the last audio track dies; rollback on refused commits
- Playback: block-start slot scan, per-sample granule reads, grid
  resync at every bar tick (sub-sample corrections, drift-proof);
  clock accumulator promoted to double (float32 drifted the grid tens
  of ms/hour against sample-exact audio - an audible flam mid-jam)
- Live button gesture: single press = grab pads+keys, DOUBLE press =
  grab guitar; the 400ms decision delay is free because the window is
  cut at the FIRST press's clock position (retrospective capture)
- Protocol: MSG_AUDIO_PEAKS, MSG_POOL, SRC_ACTIVITY grows guitar
  fields; audio rings/anchors flushed on rewind

Companion:
- Guitar history ring (amber) + Grab Guitar + per-source length
- Audio memory gauge 'N/M bars free' with tempo-lock tooltip
- Audio lanes render mirrored waveform peaks
- Track-data retry handles audio tracks via peaks presence

Verified: make (ARM/QSPI), make test (62/62 incl. 9 new pool/copy
tests), tsc + vite build. Hardware: needs the guitar.

+++ src/clock.h
@@ -52,7 +52,7 @@ class Engine
         locked_            = false;
         bpm_               = DEFAULT_BPM;
         tick_              = 0;
-        accumulator_       = 0.0f;
+        accumulator_       = 0.0;
         tap_count_         = 0;
         last_tap_ms_       = 0;
         dirty_             = true;
@@ -77,7 +77,7 @@ class Engine
         }
         for(size_t f = 0; f < nframes; f++)
         {
-            accumulator_ += 1.0f;
+            accumulator_ += 1.0;
             if(accumulator_ >= samples_per_tick_)
             {
                 accumulator_ -= samples_per_tick_;
@@ -148,7 +148,7 @@ class Engine
         playing_           = false;
         preroll_remaining_ = 0;
         tick_              = 0;
-        accumulator_       = 0.0f;
+        accumulator_       = 0.0;
         run_start_tick_    = 0;
         dirty_             = true;
     }
@@ -270,8 +270,11 @@ class Engine
   private:
     void UpdateTickInterval()
     {
-        float ticks_per_second = (bpm_ * (float)PPQN) / 60.0f;
-        samples_per_tick_      = sample_rate_ / ticks_per_second;
+        // double precision: float32 accumulation drifts the grid by tens
+        // of ms per hour against the sample-exact audio loops — enough to
+        // flam the drums against the guitar mid-jam
+        double ticks_per_second = ((double)bpm_ * (double)PPQN) / 60.0;
+        samples_per_tick_       = (double)sample_rate_ / ticks_per_second;
     }
 
     float    sample_rate_;
@@ -279,8 +282,8 @@ class Engine
     bool     locked_;
     float    bpm_;
     uint32_t tick_;
-    float    accumulator_;
-    float    samples_per_tick_;
+    double   accumulator_;
+    double   samples_per_tick_;
     bool     dirty_;
 
     uint32_t intervals_[TAP_HISTORY];
